mcp-servers/network-ops/requirements.txt
fastmcp>=2.0
mysql-connector-python>=8.0
langchain-chroma>=0.1
langchain-ollama>=0.1
langchain-openai>=0.1
pydantic>=2.0
FlagEmbedding>=1.2
pyyaml>=6.0
